# finalProject-main/finalProject-main/src/api/swagger.json
{
  "swagger": "2.0",
  "info": {
      "description": "Cocktail API",
      "version":"1.0.0",
      "title": "Cocktail API",
      "contact": {
        "name": "Cocktail",
        "url": "http://localhost:8000"
      },
      "license": {
          "name": "MIT License",
          "url": "https://opensource.org/licenses/MIT"
      }
  },
  "basePath": "/api",
  "schemes": [
  "https",
  "http"
],
"produces": [
  "application/json"
],
"consumes": [
  "application/json"
],
"paths": {
...
    "/cocktail":{
      "post": {
          "summary": "Create new cocktail",
          "description": "Create new cocktail. Auth is required",
          "tags": [
            "cocktail"
          ],
          "operationId": "CreateCocktail",
          "parameters": [
              {
                  "name": "body",
                  "in": "body",
                  "required": true,
                  "description": "Details of the new cocktail",
                  "schema": {
                    "$ref": "#/definitions/NewCocktail"
                  }
                }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/CocktailResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
      }
    },
...
},
"definitions":{
    "Cocktail":{
        "type": "object",
        "properties": {
          "name": {
              "type":"string"
          },
          "category": {
              "type":"string"
          },
          "instruction": {
              "type":"string"
          },
          "image": {
              "type":"string"
          },
          "ingredients": {
              "type":"array",
              "items":{
                  "type": "string"
              }
          }
        },
        "required": [
          "name",
          "category",
          "instruction",
          "image",
          "ingredients"
        ]
    }, 
    "NewCocktail": {
      "type": "object",
      "properties": {
          "name": {
              "type":"string"
          },
          "category": {
              "type":"string"
          },
          "instruction": {
              "type":"string"
          },
          "image": {
              "type":"string"
          },
          "ingredients": {
              "type":"array",
              "items":{
                  "type": "string"
              }
          }
      },
      "required": [
          "name",
          "category",
          "instruction",
          "image",
          "ingredients"
      ]
    }
    ,"CocktailResponse": {
      "type": "object",
      "properties": {
        "cocktail": {
          "$ref": "#/definitions/Cocktail"
        }
      },
      "required": [
        "cocktail"
      ]
    }, 
...
  "GenericErrorModel": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "object",
          "properties": {
            "body": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "body"
          ]
        }
      },
      "required": [
        "errors"
      ]
    }
}
}